Add DecryptLogistics function in ECPayLogistics
A function named DecryptLogistics is added to the ECPayLogistics service. It takes the response body as an argument and decodes it from JSON to the ECPayLogistics structure. This function replaces the previous code of manual JSON decryption. The code for redirection to logistic selection is also updated to incorporate this new function.