Remove logging statements and refactor function parameters
Logging statements were removed from helpers.go to clean up the output. Also, the function signature for DecryptLogistics in ecpay_logistics.go was changed to accept a byte slice instead of an io.ReadCloser. This makes the function easier to test and use, as one can now pass in a byte slice directly. An UpdateTempTrade function was also added.